new acl system implemented and one-to-one fixed
New acl system has been implemented in both admin cms and the reference implementation,
and one-to-one messaging has been fixed. Now the user_id is not broadcasted to clients
in the room, but a private room uuid for each user is broadcasted instead. Some other
way needs to be implemented in the future for getting the room id of individual users,
so a user can only send private messages to another user if they're actually connected,
e.g. if they're friends or however the business logic could look like. One way could be
for implemented to use a rest api from dino to get private room id's from a list of
user id's, though if in the future one of those users removes their connection or gives
the room id to another user, we still need to forbid those events/messages.